+++ private/gen/proto/api/buf/alpha/registry/v1alpha1/registryv1alpha1api/plugin_curation.pb.go
@@ -18,6 +18,7 @@ package registryv1alpha1api
 
 import (
 	context "context"
+	v1 "github.com/bufbuild/buf/private/gen/proto/go/buf/alpha/image/v1"
 	v1alpha1 "github.com/bufbuild/buf/private/gen/proto/go/buf/alpha/registry/v1alpha1"
 )
 
@@ -39,11 +40,11 @@ type PluginCurationService interface {
 		version string,
 		containerImageDigest string,
 		options []string,
-		dependencies []string,
+		dependencies []*v1alpha1.CuratedPluginReference,
 		sourceUrl string,
 		description string,
 		runtimeConfig *v1alpha1.RuntimeConfig,
-		revision int32,
+		revision uint32,
 	) (configuration *v1alpha1.CuratedPlugin, err error)
 	// GetLatestCuratedPlugin returns the latest version of a plugin matching given parameters.
 	GetLatestCuratedPlugin(
@@ -52,4 +53,16 @@ type PluginCurationService interface {
 		name string,
 		version string,
 	) (plugin *v1alpha1.CuratedPlugin, err error)
+}
+
+// CodeGenerationService generates code using remote plugins.
+type CodeGenerationService interface {
+	// GenerateCode generates code using the specified remote plugins.
+	GenerateCode(
+		ctx context.Context,
+		image *v1.Image,
+		requests []*v1alpha1.PluginGenerationRequest,
+		includeImports bool,
+		includeWellKnownTypes bool,
+	) (responses []*v1alpha1.PluginGenerationResponse, err error)
 }